Updated tag loop
Swapped checking to see if the grabbed tags were in the JSON object to checking each JSON object expected tag is in the list of grabbed tags from the page. This distinction seems small, but the JSON object is what I'm looking for and making sure is present, I don't really care what is on the page as long as the listed tags in the JSON object are present.